--- src/run.py
@@ -1,30 +1,0 @@
-from flask import Flask, request, redirect
-import twilio.twiml
- 
-app = Flask(__name__)
- 
-# Try adding your own number to this list!
-callers = {
-    "+16507769918": "Gavin Chan",
-    "+14158675310": "Boots",
-    "+14158675311": "Virgil",
-    "+12566671171": "MOSC",
-}
- 
-@app.route("/", methods=['GET', 'POST'])
-def hello_monkey():
-    # Get the caller's phone number from the incoming Twilio request
-    from_number = request.values.get('From', None)
-    resp = twilio.twiml.Response()
- 
-    # if the caller is someone we know:
-    if from_number in callers:
-        # Greet the caller by name
-        resp.say("Hello " + callers[from_number])
-    else:
-        resp.say("Hello Monkey")
- 
-    return str(resp)
- 
-if __name__ == "__main__":
-    app.run(debug=True)
