+++ src/templates/503view.html
@@ -1,0 +1,5 @@
+{% block title %}503: Service Unavailable{% endblock %}
+{% block body %}
+  <h1>Generated error code 503</h1>
+  <p>Custom template 503view.html</p>
+{% endblock %}
