+++ src/templates/hello_world.html
@@ -1,0 +1,5 @@
+{% block title %}Hello World{% endblock %}
+{% block body %}
+  <h1>This is the root landing page</h1>
+  <a href="view-503">Go to 503 page</a>
+{% endblock %}
